requirements.txt
fastapi[all]
psycopg2-binary
sqlalchemy
databases
asyncpg
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings